requirements.txt
# ===================
# Core Framework
# ===================
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# ===================
# LangChain Ecosystem
# ===================
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0

# ===================
# LLM Routing
# ===================
litellm>=1.30.0

# ===================
# Audio Processing
# ===================
openai>=1.40.0,<1.60.0
pydub>=0.25.1

# ===================
# Visualization
# ===================
streamlit-flow-component>=1.6.0

# ===================
# Utilities
# ===================
httpx>=0.27.0
tenacity>=8.2.0
redis>=5.0.0

# ===================
# Development & Testing
# ===================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
